move configs to base package
